+++ docs/components/SignatureRequest/index.tsx
@@ -1,28 +1,32 @@
 import { type ComponentType, useEffect, useState } from 'react'
-import '../styles.css'
+import '../../styles.css'
 
+/**
+ * Lazy-loading wrapper that avoids SSR issues with wagmi/wallet code.
+ * The actual demo is in Demo.tsx and must be a separate module so Vite
+ * resolves its wagmi imports to the same instance as WalletProvider.
+ */
 export default function SignatureRequestExample() {
   const [Component, setComponent] = useState<ComponentType | null>(null)
 
   useEffect(() => {
-    Promise.all([
-      import('./WalletProvider'),
-      import('./SignatureRequestDemo'),
-    ]).then(([providerMod, demoMod]) => {
-      const Provider = providerMod.WalletProvider
-      const Demo = demoMod.default
+    Promise.all([import('../WalletProvider'), import('./Demo')]).then(
+      ([providerMod, demoMod]) => {
+        const Provider = providerMod.WalletProvider
+        const Demo = demoMod.default
 
-      setComponent(
-        () =>
-          function Wrapped() {
-            return (
-              <Provider>
-                <Demo />
-              </Provider>
-            )
-          },
-      )
-    })
+        setComponent(
+          () =>
+            function Wrapped() {
+              return (
+                <Provider>
+                  <Demo />
+                </Provider>
+              )
+            },
+        )
+      },
+    )
   }, [])
 
   if (!Component) {
