+++ docs/pages/react/features/transaction-signing.mdx
@@ -1,4 +1,4 @@
-import SignatureRequestExample from '../../../components/SignatureRequestExample'
+import SignatureRequestExample from '../../../components/SignatureRequest'
 
 # Transaction Signing
